requirements.txt
flask
sqlalchemy
pandas
python-dotenv
bcrypt
pyjwt
streamlit
plotly
matplotlib
seaborn
requests